DynamicalBilliards.MushroomTools — Module

MushroomTools

Module containing many functions helpful in simulating (perfect) mushroom billiards, see billiard_mushroom. Contains stuff like initializing efficiently regular or chaotic particles and functions that return the corresponding chaotic or regular phase-space volumes or portions. The functions V_3D_tot and V_3D_reg use equations derived in ref. [1].

Made by Lukas Hupe.

References

[1] A. Barnett & T. Betcke, Chaos 17, 043125 (2007).

Notice that the name MushroomTools is exported by DynamicalBilliards. The functions within it are not though, so you have to access them like e.g. MushroomTools.randomchaotic.

DynamicalBilliards.MushroomTools.V_2D_tot — Method

V_2D_tot(l,w,r)

Return the total boundary map volume (2D) of a billiard_mushroom parameterized by (l,w,r).

DynamicalBilliards.MushroomTools.V_3D_tot — Method

V_3D_tot(l,w,r)

Return the total phasespace volume (3D) of a billiard_mushroom parameterized by (l,w,r).

DynamicalBilliards.MushroomTools.g_c_2D — Method

g_c_2D(l, w, r)

Return the chaotic phasespace portion of the boundary map (2D) of a billiard_mushroom with stem length l, stem width w and cap radious r. This result is known analytically, see MushroomTools for references.

DynamicalBilliards.MushroomTools.g_c_3D — Method

g_c_3D(l, w, r)

Return the chaotic phasespace portion of the full (3D) phase-space of a billiard_mushroom with stem length l, stem width w and cap radious r. This result is known analytically, see MushroomTools for references.

DynamicalBilliards.MushroomTools.g_r_2D — Method

g_r_2D(l, w, r)

Return the regular phasespace portion of the boundary map (2D) of a billiard_mushroom with stem length l, stem width w and cap radious r. This result is known analytically, see MushroomTools for references.

DynamicalBilliards.MushroomTools.g_r_3D — Method

g_r_3D(l, w, r)

Return the regular phasespace portion of the full (3D) phase-space of a billiard_mushroom with stem length l, stem width w and cap radious r. This result is known analytically, see MushroomTools for references.

DynamicalBilliards.MushroomTools.randin_mushroom — Method

randin_mushroom(l, w, r [, ω])

Generate a random particle within the billiard_mushroom parameterised by l, w and r. If ω is given the particle is magnetic instead.

This function is much more efficient than randominside.

DynamicalBilliards.MushroomTools.randomchaotic — Method

randomchaotic(l, w, r)

Generate a chaotic particle, i.e. not trapped in the cap.

DynamicalBilliards.MushroomTools.randomregular — Method

randomregular(l, w, r)

Generate a regular particle (i.e. trapped in the cap).